refactor  solana
