+++ packages/huma-shared/src/solana/const.ts
@@ -1,3 +1,3 @@
 import { BN } from '@coral-xyz/anchor'
 
-export const BP_FACTOR = BN.from(10000)
+export const SOLANA_BP_FACTOR = BN.from(10000)
