+++ packages/huma-shared/src/solana/idl/localhost.ts
