+++ packages/huma-shared/src/solana/index.ts
@@ -1,4 +1,4 @@
-export * from './programHelpers'
-export * from './tokenAccountHelpers'
-export * from './poolHelper'
+export * from './utils'
 export * from './const'
+export * from './idl'
+export * from './pool'
