+++ packages/huma-shared/src/v2/index.ts
@@ -1,4 +1,3 @@
 export * from './types'
 export * from './abis'
-export * from './idl'
 export * from './utils'
