+++ packages/huma-shared/src/v2/utils/index.ts
@@ -4,5 +4,3 @@ export * from './firstLossCoverContract'
 export * from './poolConfigContract'
 export * from './poolContract'
 export * from './trancheVaultContract'
-export * from './solanaMetadataFind'
-export * from './solanaTokenDecimals'
